refactor(runtime): move event bus from daemon and migrate to Storage
EventBus now lives in crabtalk-runtime and persists its
subscriptions at the `events/subscriptions.toml` key in the runtime
Storage instead of a hard-coded file path. The bus takes a generic
`fire` callback at construction — it has no knowledge of
DaemonEvent or ClientMessage — and the daemon wires up a closure
that forwards matched subscriptions into its own event loop as
DaemonEvent::Message. Same behaviour at runtime, just no reach-
through into daemon types from the bus itself.

The daemon's event_bus.rs is deleted; all callers import from
runtime::event_bus (EventBus + EventSubscription + FireCallback +
SUBSCRIPTIONS_KEY). builder.rs constructs the bus with
runtime.storage.clone() and the fire closure. toml gains a spot in
crabtalk-runtime's Cargo.toml for the subscription blob (de)serde.
