+++ crates/daemon/src/lib.rs
@@ -4,7 +4,6 @@
 pub mod config;
 pub mod cron;
 pub mod daemon;
-pub mod event_bus;
 pub mod hook;
 pub mod provider;
 pub mod storage;
