+++ crates/runtime/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod ask_user;
 pub mod config;
 pub mod env;
+pub mod event_bus;
 pub mod host;
 pub mod mcp;
 pub mod memory;
